pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Quiz service REST API"
authors = [""]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.4"
djangorestframework = "^3.14.0"
markdown = "^3.4.1"
django-filter = "^22.1"
mysqlclient = "^2.1.1"
tzdata = "^2022.7"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^7.0.1"}
djangorestframework-simplejwt = "^5.2.2"
django-redis = "^5.2.0"
django-cors-headers = "^3.13.0"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
